Move github/client to github and github/release to release
