+++ pkg/github/client.go
@@ -1,4 +1,4 @@
-package client
+package github
 
 import (
 	"context"
