+++ pkg/github/client_iface.go
@@ -1,4 +1,4 @@
-package client
+package github
 
 import (
 	"github.com/google/go-github/github"
